refactor: :recycle: refactored order routes
